+++ src/app/app.component.html
@@ -1,4 +1,4 @@
-<app-header></app-header>
+<app-header></app-header> 
 <router-outlet></router-outlet>
 
 <app-contact></app-contact>
