+++ src/app/auth/login/login.component.css
@@ -1,17 +1,3 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: Arial, sans-serif;
-    line-height: 1.6;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-}
-
 .container {
     max-width: 1200px;
     margin: 0 auto;
@@ -96,7 +82,7 @@ h1 {
 .btn-login {
     width: 100%;
     background-color: #E7AF3E; 
-    color: black;
+    color: #fff;
     font-size: 1.1em;
     padding: 15px;
     margin-top: 20px;
