+++ src/app/auth/login/login.component.html
@@ -8,6 +8,7 @@
 </head>
 <body>
 
+    <!-- Main Content -->
     <main class="main-content">
         <h1>Inicia sesión</h1>
         
@@ -27,7 +28,7 @@
                     <input type="password" id="contrasena" name="contrasena" formControlName="contrasena" required>
                 </div>
                 
-                <button type="submit" class="btn btn-login" routerLink="/createevent" (click)="navigateTo('/createevent')" >INICIAR</button>
+                <button type="submit" class="btn btn-login">INICIAR</button>
             </form>
         </div>
     </main>
