+++ src/app/components/cards-org/cards-org.component.html
@@ -1,0 +1,60 @@
+<main class="main">
+  <div class="container">
+    <link rel="stylesheet" href="cards-org.component.css">
+    <div class="title-section">
+      <h1>Organizaciones Benéficas</h1>
+      <!-- Modificar el botón para usarrouterLink-->
+      <button class="btn btn-login" routerLink="'/login'">Inicia sesión</button>
+    </div>
+
+    <div class="organizations-grid">
+      <div class="org-card">
+        <img src="imagen3.png" alt="Manos Unidas Logo" class="org-logo">
+        <h2 class="org-name">Nombre</h2>
+        <p class="org-description">
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
+        </p>
+        <div class="contact-icons">
+          <a href="tel:" class="icon icon-phone"><img src="icono4.png" alt=""></a>
+          <a href="mailto:" class="icon icon-email"><img src="icono5.png" alt=""></a>
+        </div>
+      </div>
+
+      <div class="org-card">
+        <img src="imagen3.png" alt="Manos Unidas Logo" class="org-logo">
+        <h2 class="org-name">Nombre</h2>
+        <p class="org-description">
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
+        </p>
+        <div class="contact-icons">
+          <a href="tel:" class="icon icon-phone"><img src="icono4.png" alt=""></a>
+          <a href="mailto:" class="icon icon-email"><img src="icono5.png" alt=""></a>
+        </div>
+      </div>
+
+      <div class="org-card">
+        <img src="imagen3.png" alt="Manos Unidas Logo" class="org-logo">
+        <h2 class="org-name">Nombre</h2>
+        <p class="org-description">
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
+        </p>
+        <div class="contact-icons">
+          <a href="tel:" class="icon icon-phone"><img src="icono4.png" alt=""></a>
+          <a href="mailto:" class="icon icon-email"><img src="icono5.png" alt=""></a>
+        </div>
+      </div>
+    </div>
+
+    <div class="join-section">
+      <div class="join-image">
+        <img src="imagen4.png" alt="Equipo de voluntarios">
+      </div>
+      <div class="join-content">
+        <h2>¿Eres una organización?</h2>
+        <h3>¡ÚNETE AHORA!</h3>
+        <!-- Modificar el botón para usarrouterLink-->
+        <button class="btn btn-join" routerLink="'/formorg'">UNIRME</button>
+      </div>
+    </div>
+  </div>
+</main>
