+++ src/app/components/cards-org/cards-org.component.ts
@@ -1,5 +1,4 @@
 import { Component } from '@angular/core';
-import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-cards-org',
@@ -7,10 +6,6 @@ import { Router } from '@angular/router';
   styleUrls: ['./cards-org.component.css']
 })
 export class CardsOrgComponent {
-  constructor(private router: Router) {}
 
-  navigateTo(route: string) {
-    this.router.navigateByUrl(route);
-    return false; 
-  }
+ 
 }
