+++ src/app/components/contact/contact.component.html
@@ -1,11 +1,11 @@
 <div class="container contact-content">
-    <div class="contact-info">
-      <img src="logo.png" alt="ONAT Logo" class="logo">
-      <link rel="stylesheet" href="contact.component.css">
-      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
-    </div>
-    <div class="contact-form">
-      <h2>Contáctanos</h2>
+  <div class="contact-info">
+    <img src="logo.png" alt="ONAT Logo" class="logo">
+    <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
+  </div>
+  <div class="contact-form">
+    <h2>Contáctanos</h2>
+    <div class="contact-details">
       <div class="contact-item">
         <span class="icon"><img src="icono4.png" alt=""></span>
         <span>(+52) 9612198101</span>
@@ -14,7 +14,9 @@
         <span class="icon"><img src="icono5.png" alt=""></span>
         <span>onatgmail.com</span>
       </div>
-      <button class="btn btn-donate" routerLink="/directdonation">Donar!</button>
+    </div>
+    <div class="contact-button-wrapper">
+      <button class="btn btn-donate" href="/directdonation" (click)="navigateTo('/directdonation')">Donar!</button>
     </div>
   </div>
-  
+</div>
