+++ src/app/components/contact/contact.component.ts
@@ -1,10 +1,16 @@
 import { Component } from '@angular/core';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-contact',
   templateUrl: './contact.component.html',
-  styleUrl: './contact.component.css'
+  styleUrls: ['./contact.component.css']
 })
 export class ContactComponent {
+  constructor(private router: Router) {}
 
+  navigateTo(route: string) {
+    this.router.navigateByUrl(route);
+    return false; 
+  }
 }
