+++ src/app/components/header/header.component.html
@@ -1,20 +1,20 @@
 <nav class="navbar">
   <div class="container-nav">
-    <a class="navbar-brand" href="/">
+    <a class="navbar-brand" routerLink="/">
       <img src="logo.png" alt="ONAT Logo" class="logo" />
     </a>
     <ul class="navbar-nav">
       <li class="nav-item">
-        <a class="nav-link" href="/cardsorg">Organizaciones</a>
+        <a class="nav-link" routerLink="/cardsorg">Organizaciones</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" href="#">Eventos</a>
+        <a class="nav-link" routerLink="/eventos">Eventos</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link" href="#">Membresía</a>
+        <a class="nav-link" routerLink="/membresia">Membresía</a>
       </li>
       <li class="nav-item">
-        <a class="nav-link nav-link-button" href="#">Donar</a>
+        <a class="nav-link nav-link-button" routerLink="/donar">Donar</a>
       </li>
     </ul>
   </div>
