+++ src/app/donation/membership-donation/membership-donation.component.css
@@ -1,17 +1,3 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: Arial, sans-serif;
-    line-height: 1.6;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    background-color: #ffffff; 
-}
 
 .container {
     max-width: 1200px;
