--- src/app/event/create-event/create-event.component.css
@@ -1,254 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: Arial, sans-serif;
-    line-height: 1.6;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    background-color: #B2D5F1; 
-}
-
-.container {
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 0 20px;
-}
-
-.btn {
-    padding: 12px 24px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    font-weight: bold;
-    transition: background-color 0.3s ease;
-}
-
-.btn-donate {
-    background-color: #ff7f50; 
-    color: #ffffff;
-}
-
-.btn-donate:hover {
-    background-color: #E76F51;
-}
-
-/* Main Content Styles */
-.main-content {
-    flex: 1;
-    padding: 40px 0;
-}
-
-h1, h2 {
-    text-align: center;
-    color: #003366; 
-}
-
-h1 {
-    font-size: 2em;
-    margin-bottom: 10px;
-}
-
-h2 {
-    font-size: 1.8em;
-    margin-bottom: 30px;
-}
-
-/* Form Styles */
-.event-form {
-    background-color: #E9F3FB; 
-    padding: 30px;
-    border-radius: 10px;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-    margin-bottom: 40px;
-}
-
-.form-grid {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 30px;
-}
-
-.form-group {
-    margin-bottom: 20px;
-}
-
-.form-group label {
-    display: block;
-    margin-bottom: 5px;
-    color: #003366;
-    font-weight: 500;
-}
-
-.form-group input,
-.form-group textarea {
-    width: 100%;
-    padding: 12px;
-    border: 1px solid #ddd;
-    border-radius: 5px;
-    font-size: 16px;
-    background-color: #f5f5f5;
-}
-
-.form-group textarea {
-    height: 100px;
-    resize: vertical;
-}
-
-.input-icon {
-    position: relative;
-}
-
-.icon {
-    position: absolute;
-    right: 12px;
-    top: 50%;
-    transform: translateY(-50%);
-    width: 20px;
-    height: 20px;
-    background-size: contain;
-    background-repeat: no-repeat;
-}
-
-.calendar-icon {
-    background-image: url('/public/icono7.png');
-}
-
-.clock-icon {
-    background-image: url('/public/icono8.png');
-}
-
-/* Upload Area Styles */
-.upload-area {
-    width: 100%;
-    height: 200px;
-    border: 2px dashed #ddd;
-    border-radius: 5px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-    background-color: #f5f5f5;
-}
-
-.upload-icon {
-    width: 48px;
-    height: 48px;
-}
-
-/* Checkbox Styles */
-.checkbox-grid {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 15px;
-    margin-top: 10px;
-}
-
-.checkbox-label {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    cursor: pointer;
-}
-
-.checkbox-label input[type="checkbox"] {
-    width: 18px;
-    height: 18px;
-}
-
-/* Button Styles */
-.btn-create {
-    width: 100%;
-    background-color: #E7AF3E; 
-    color: black; 
-    font-size: 1.1em;
-    padding: 15px;
-    margin-top: 20px;
-}
-
-/* History Section Styles */
-.history-section {
-    background-color: #B2D5F1; 
-    border-radius: 10px;
-    overflow: hidden;
-    margin-top: 40px;
-
-}
-
-.history-content {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 30px;
-    align-items: center;
-}
-
-.history-image {
-    width: 400px;
-    height: 400px;
-    object-fit: cover;
-}
-
-.history-text {
-    padding: 30px;
-    text-align: center;
-}
-
-.btn-history {
-    background-color: #E7AF3E; 
-    color: black; 
-    display: inline-flex;
-    align-items: center;
-    gap: 10px;
-    margin-top: 20px;
-}
-
-.lock-icon {
-    width: 20px;
-    height: 20px;
-    background-image: url('/icono9.png');
-    background-size: contain;
-    background-repeat: no-repeat;
-}
-
-/* Responsive Styles */
-@media (max-width: 768px) {
-    .nav {
-        display: none;
-    }
-    
-    .form-grid {
-        grid-template-columns: 1fr;
-    }
-    
-    .history-content {
-        grid-template-columns: 1fr;
-    }
-    
-    .checkbox-grid {
-        grid-template-columns: 1fr;
-    }
-}
-
-/* Input States */
-input:focus,
-textarea:focus {
-    outline: none;
-    border-color: #007bb5; 
-    box-shadow: 0 0 0 2px rgba(119, 184, 232, 0.2);
-}
-
-/* Upload Area Hover State */
-.upload-area:hover {
-    border-color: #007bb5; 
-    background-color: #fafafa;
-}
-
-/* Button Hover States */
-.btn-create:hover,
-.btn-history:hover {
-    background-color: #e4a62a;
-}
