--- src/app/event/create-event/create-event.component.html
@@ -1,99 +1,0 @@
-    <main class="main-content">
-        <link rel="stylesheet" href="create-event.component.css">
-        <div class="container">
-            <h1>Crea un evento</h1>
-            <h2>¡CREA TU EVENTO AHORA!</h2>
-
-            <form class="event-form">
-                <div class="form-grid">
-                    <div class="form-column">
-                        <div class="form-group">
-                            <label for="nombre">Nombre del evento:</label>
-                            <input type="text" id="nombre" name="nombre" required>
-                        </div>
-
-                        <div class="form-group">
-                            <label for="direccion">Dirección:</label>
-                            <input type="text" id="direccion" name="direccion" required>
-                        </div>
-
-                        <div class="form-group">
-                            <label for="fecha">Fecha:</label>
-                            <div class="input-icon">
-                                <input type="date" id="fecha" name="fecha" required>
-                                <span class="icon calendar-icon"></span>
-                            </div>
-                        </div>
-
-                        <div class="form-group">
-                            <label for="hora-inicio">Hora de inicio:</label>
-                            <div class="input-icon">
-                                <input type="time" id="hora-inicio" name="hora-inicio" required>
-                                <span class="icon clock-icon"></span>
-                            </div>
-                        </div>
-
-                        <div class="form-group">
-                            <label for="hora-termino">Hora de termino:</label>
-                            <div class="input-icon">
-                                <input type="time" id="hora-termino" name="hora-termino" required>
-                                <span class="icon clock-icon"></span>
-                            </div>
-                        </div>
-
-                        <div class="form-group">
-                            <label for="descripcion">Descripción:</label>
-                            <textarea id="descripcion" name="descripcion" required></textarea>
-                        </div>
-                    </div>
-
-                    <div class="form-column">
-                        <div class="form-group">
-                            <label>Sube una imagen no mayor a 1GB</label>
-                            <div class="upload-area">
-                                <img src="icono6.png" alt="Upload" class="upload-icon">
-                            </div>
-                        </div>
-
-                        <div class="form-group">
-                            <label>Selecciona los productos que puedes recibir:</label>
-                            <div class="checkbox-grid">
-                                <label class="checkbox-label">
-                                    <input type="checkbox" name="productos" value="atun">
-                                    <span>Atún</span>
-                                </label>
-                                <label class="checkbox-label">
-                                    <input type="checkbox" name="productos" value="frijol">
-                                    <span>Frijol</span>
-                                </label>
-                                <label class="checkbox-label">
-                                    <input type="checkbox" name="productos" value="arroz">
-                                    <span>Arroz</span>
-                                </label>
-                                <label class="checkbox-label">
-                                    <input type="checkbox" name="productos" value="sopas">
-                                    <span>Sopas/Pasta</span>
-                                </label>
-                            </div>
-                        </div>
-
-                        <button type="submit" class="btn btn-create">CREAR</button>
-                    </div>
-                </div>
-            </form>
-
-            <div class="history-section">
-                <div class="history-content">
-                    <img src="imagen5.png" alt="Voluntarios con donaciones" class="history-image">
-                    <div class="history-text">
-                        <h2>¿Quieres ver quien te dono?</h2>
-                        <button class="btn btn-history" (click)="navigateTo('/historydonation')">
-                            VER HISTORIAL
-                            <span class="lock-icon"></span>
-                        </button>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </main>
-
