--- src/app/event/event-donation/event-donation.component.css
@@ -1,206 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: Arial, sans-serif;
-    line-height: 1.6;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-}
-
-.container {
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 0 20px;
-}
-
-.btn {
-    padding: 12px 24px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    font-weight: bold;
-    transition: background-color 0.3s ease;
-}
-
-.btn-donate {
-    background-color: #ff7f50; 
-    color: #ffffff;
-}
-
-.btn-donate:hover {
-    background-color: #E76F51;
-}
-
-/* Main Content Styles */
-.main-content {
-    flex: 1;
-    padding: 40px 0;
-}
-
-h1 {
-    text-align: center;
-    color: #003366; 
-    font-size: 2em;
-    margin-bottom: 40px;
-}
-
-.content-grid {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 40px;
-}
-
-/* Organization Info Styles */
-.org-info {
-    padding-right: 20px;
-}
-
-.org-logo {
-    width: 200px;
-    margin-bottom: 20px;
-}
-
-.org-description {
-    color: #7d7d7d; 
-    font-size: 0.9em;
-    line-height: 1.8;
-}
-
-/* Donation Form Styles */
-.donation-form {
-    background: #ffffff; 
-    padding: 20px;
-    border-radius: 10px;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-}
-
-.form-row {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 20px;
-}
-
-.form-group {
-    margin-bottom: 20px;
-}
-
-.form-group label {
-    display: block;
-    margin-bottom: 5px;
-    color: #003366; 
-    font-weight: 500;
-}
-
-.form-group input,
-.form-group select {
-    width: 100%;
-    padding: 10px;
-    border: 1px solid #ddd;
-    border-radius: 5px;
-    font-size: 16px;
-}
-
-.select-wrapper {
-    position: relative;
-}
-
-.select-wrapper::after {
-    position: absolute;
-    right: 10px;
-    top: 50%;
-    transform: translateY(-50%);
-    pointer-events: none;
-    font-size: 12px;
-    color: #7d7d7d; 
-}
-
-/* Products Table Styles */
-.products-table {
-    width: 100%;
-    border-collapse: collapse;
-    margin-bottom: 10px;
-}
-
-.products-table th,
-.products-table td {
-    padding: 10px;
-    text-align: left;
-    border-bottom: 1px solid #ddd;
-}
-
-.products-table th {
-    font-weight: 500;
-    color: #003366; 
-}
-
-.quantity-select {
-    width: 100px;
-    padding: 5px;
-    border: 1px solid #ddd;
-    border-radius: 5px;
-}
-
-.form-note {
-    font-size: 0.8em;
-    color: #7d7d7d; 
-    margin-top: 10px;
-}
-
-.btn-submit {
-    width: 100%;
-    background-color: #E7AF3E; 
-    color: black; 
-    padding: 15px;
-    font-size: 1.1em;
-    margin-top: 20px;
-}
-
-/* Responsive Styles */
-@media (max-width: 768px) {
-    .nav {
-        display: none;
-    }
-    
-    .content-grid {
-        grid-template-columns: 1fr;
-    }
-    
-    .form-row {
-        grid-template-columns: 1fr;
-    }
-    
-    .footer-content {
-        grid-template-columns: 1fr;
-        text-align: center;
-    }
-    
-    .footer-right {
-        justify-content: center;
-    }
-    
-    .contact-info {
-        align-items: center;
-    }
-    
-    .contact-item {
-        justify-content: center;
-    }
-}
-
-/* Input States */
-input:focus,
-select:focus {
-    outline: none;
-    border-color: #77b8e8; /* primary color */
-    box-shadow: 0 0 0 2px rgba(119, 184, 232, 0.2);
-}
-
-/* Button Hover States */
-.btn-submit:hover {
-    background-color: #ebaa2a;
-}
