--- src/app/event/event-donation/event-donation.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-event-donation',
-  templateUrl: './event-donation.component.html',
-  styleUrl: './event-donation.component.css'
-})
-export class EventDonationComponent {
-
-}
