--- src/app/event/event.module.ts
@@ -1,22 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { CreateEventComponent } from './create-event/create-event.component';
-import { EventsComponent } from './events/events.component';
-import { EventDonationComponent } from './event-donation/event-donation.component';
-
-@NgModule({
-  declarations: [
-    CreateEventComponent,
-    EventsComponent,
-    EventDonationComponent
-  ],
-  imports: [
-    CommonModule
-  ],
-  exports: [
-    CreateEventComponent,
-    EventsComponent,
-    EventDonationComponent
-  ]
-})
-export class EventModule { }
