--- src/app/event/events/events.component.css
@@ -1,135 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: Arial, sans-serif;
-    line-height: 1.6;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-}
-
-.container {
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 0 20px;
-}
-
-.btn {
-    padding: 12px 24px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    font-weight: bold;
-    transition: background-color 0.3s ease;
-}
-
-.btn-donate {
-    background-color: #E7AF3E; 
-    color: black; 
-}
-
-.btn-donate:hover {
-    background-color: #e7a82a;
-}
-
-/* Main Content Styles */
-.main-content {
-    flex: 1;
-    padding: 40px 0;
-}
-
-h1 {
-    text-align: center;
-    color: #003366; 
-    font-size: 2.5em;
-    margin-bottom: 40px;
-}
-
-/* Events Grid */
-.events-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-    gap: 30px;
-    margin-bottom: 40px;
-}
-
-.event-card {
-    background: #ffffff; 
-    border-radius: 10px;
-    padding: 20px;
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-    text-align: center;
-    transition: transform 0.3s ease;
-}
-
-.event-card:hover {
-    transform: translateY(-5px);
-}
-
-.event-datetime {
-    color: #7d7d7d; 
-    font-size: 0.9em;
-    margin-bottom: 15px;
-}
-
-.event-logo {
-    width: 150px;
-    height: 150px;
-    margin: 0 auto 20px;
-}
-
-.event-title {
-    font-size: 1.5em;
-    color: #003366; 
-    margin-bottom: 10px;
-}
-
-.event-address,
-.event-organization {
-    color: #7d7d7d; 
-    margin-bottom: 10px;
-}
-
-.event-card .btn-donate {
-    width: 100%;
-    margin-top: 20px;
-}
-
-/* Responsive Styles */
-@media (max-width: 768px) {
-    .nav {
-        display: none;
-    }
-    
-    .events-grid {
-        grid-template-columns: 1fr;
-    }
-    
-    .footer-content {
-        grid-template-columns: 1fr;
-        text-align: center;
-    }
-    
-    .footer-right {
-        justify-content: center;
-    }
-    
-    .contact-info {
-        align-items: center;
-    }
-    
-    .contact-item {
-        justify-content: center;
-    }
-}
-
-/* Button and Card Hover States */
-.btn:hover {
-    opacity: 0.9;
-}
-
-
