--- src/app/event/events/events.component.html
@@ -1,38 +1,0 @@
-    <main class="main-content">
-        <link rel="stylesheet" href="events.component.css">
-        <div class="container">
-          <h1>Eventos</h1>
-
-          <div class="events-grid">
-              
-              <div class="event-card">
-                  <div class="event-datetime">10/12/24 01:59</div>
-                  <img src="imagen3.png" alt="Manos Unidas Logo" class="event-logo">
-                  <h2 class="event-title">Alimento Digno</h2>
-                  <p class="event-address">Av. Madero Mz.10 Lt 3</p>
-                  <p class="event-organization">Manos Unidas</p>
-                  <button class="btn btn-donate" href="/eventdonation" (click)="navigateTo('/eventdonation')">DONAR</button>
-                </div>
-
-              
-              <div class="event-card">
-                  <div class="event-datetime">10/12/24 01:59</div>
-                  <img src="imagen3.png" alt="Manos Unidas Logo" class="event-logo">
-                  <h2 class="event-title">Alimento Digno</h2>
-                  <p class="event-address">Av. Madero Mz.10 Lt 3</p>
-                  <p class="event-organization">Manos Unidas</p>
-                  <button class="btn btn-donate" href="/eventdonation" (click)="navigateTo('/eventdonation')">DONAR</button>
-                </div>
-
-              
-              <div class="event-card">
-                  <div class="event-datetime">10/12/24 01:59</div>
-                  <img src="imagen3.png" alt="Manos Unidas Logo" class="event-logo">
-                  <h2 class="event-title">Alimento Digno</h2>
-                  <p class="event-address">Av. Madero Mz.10 Lt 3</p>
-                  <p class="event-organization">Manos Unidas</p>
-                  <button class="btn btn-donate" href="/eventdonation" (click)="navigateTo('/eventdonation')">DONAR</button>
-                </div>
-          </div>
-      </div>
-  </main>
