+++ src/app/events-page/events/events.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-events',
+  templateUrl: './events.component.html',
+  styleUrl: './events.component.css'
+})
+export class EventsComponent {
+
+}
