+++ src/app/organizaciones/cards-org/cards-org.component.css
@@ -1,13 +1,3 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: Arial, sans-serif;
-    line-height: 1.6;
-}
 
 .container {
     max-width: 1200px;
