--- src/app/shared/contact/contact.component.html
@@ -1,22 +1,0 @@
-<div class="container contact-content">
-  <div class="contact-info">
-    <img src="logo.png" alt="ONAT Logo" class="logo">
-    <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
-  </div>
-  <div class="contact-form">
-    <h2>Contáctanos</h2>
-    <div class="contact-details">
-      <div class="contact-item">
-        <span class="icon"><img src="icono4.png" alt=""></span>
-        <span>(+52) 9612198101</span>
-      </div>
-      <div class="contact-item">
-        <span class="icon"><img src="icono5.png" alt=""></span>
-        <span>onatgmail.com</span>
-      </div>
-    </div>
-    <div class="contact-button-wrapper">
-      <button class="btn btn-donate" href="/directdonation" (click)="navigateTo('/directdonation')">Donar!</button>
-    </div>
-  </div>
-</div>
