--- src/app/shared/contact/contact.component.ts
@@ -1,16 +1,0 @@
-import { Component } from '@angular/core';
-import { Router } from '@angular/router';
-
-@Component({
-  selector: 'app-contact',
-  templateUrl: './contact.component.html',
-  styleUrls: ['./contact.component.css']
-})
-export class ContactComponent {
-  constructor(private router: Router) {}
-
-  navigateTo(route: string) {
-    this.router.navigateByUrl(route);
-    return false; 
-  }
-}
