--- src/app/shared/footer/footer.component.css
@@ -1,32 +1,0 @@
-
-  
-  .container {
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 0 20px;
-  }
-  
-  .btn {
-    padding: 10px 20px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    font-weight: bold;
-    transition: background-color 0.3s ease;
-  }
-  
-  .btn-donate {
-    background-color: #E5863A;
-    color: white;
-  }
-  
-  .btn-donate:hover {
-    background-color: #e4a62a;
-  }
-
-  .footer {
-    background-color: #00172F;
-    color: white;
-    padding: 10px 0;
-    text-align: center;
-  }
