--- src/app/shared/footer/footer.component.html
@@ -1,6 +1,0 @@
-<div class="footer">
-    <p>Derechos reservados</p>
-    <p>ONAT (Organización Nutriendo A Todos)</p>
-    <p>Chiapas, México.</p>
-  </div>
-
