--- src/app/shared/header/header.component.css
@@ -1,63 +1,0 @@
-body {
-    font-family: Arial, sans-serif;
-    margin: 0;
-    padding: 0;
-    color: #333;
-  }
-  
-  .container {
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 0 20px;
-  }
-  
-  .btn {
-    padding: 10px 20px;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    font-weight: bold;
-    transition: background-color 0.3s ease;
-  }
-  
-  .btn-donate {
-    background-color: #E5863A;
-    color: white;
-  }
-  
-  .btn-donate:hover {
-    background-color: #e4a62a;
-  }
-
-.container.header-content {
-    background-color: #5DB0E6;
-  }
-  
-  .header-content {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-  
-  .logo {
-    width: 80px;
-    height: 80px;
-    border-radius: 50%;
-    padding: 5px;
-  }
-  
-  .main-nav {
-    display: flex;
-    gap: 20px;
-    align-items: center;
-  }
-  
-  .nav-link {
-    color:white;
-    text-decoration: none;
-    font-weight: bold;
-  }
-  
-  .nav-link:hover {
-    text-decoration: underline;
-  }
