--- src/app/shared/header/header.component.html
@@ -1,10 +1,0 @@
-<div class="container header-content">
-  <img src="logo.png" alt="ONAT Logo" class="logo" href="/" (click)="navigateTo('/')">
-  <link rel="stylesheet" href="header.component.css">
-  <nav class="main-nav">
-    <a href="/cardsorg" class="nav-link" (click)="navigateTo('/cardsorg')">Organizaciones</a>
-    <a href="/events" class="nav-link" (click)="navigateTo('/events')">Eventos</a>
-    <a href="/membership" class="nav-link" (click)="navigateTo('/membership')">Membresía</a>
-    <button class="btn btn-donate" href="/directdonation" (click)="navigateTo('/directdonation')">Donar!</button>
-  </nav>
-</div>
