--- src/app/shared/header/header.component.ts
@@ -1,16 +1,0 @@
-import { Component } from '@angular/core';
-import { Router } from '@angular/router';
-
-@Component({
-  selector: 'app-header',
-  templateUrl: './header.component.html',
-  styleUrls: ['./header.component.css']
-})
-export class HeaderComponent {
-  constructor(private router: Router) {}
-
-  navigateTo(route: string) {
-    this.router.navigateByUrl(route);
-    return false; 
-  }
-}
